General cleanup and improvements

+++ r5dev/tier1/IConVar.h
@@ -147,7 +147,7 @@ public:
 	struct CVValue_t
 	{
 		const char* m_pszString;
-		int64_t     m_iStringLength;
+		size_t      m_iStringLength;
 		float       m_fValue;
 		int         m_nValue;
 	};
